improve new table for multiple list
