+++ explorer/src/components/Extrinsic/ExtrinsicList.tsx
@@ -2,31 +2,56 @@
 
 import { PAGE_SIZE, searchTypes } from '@/constants'
 import { numberWithCommas } from '@/utils/number'
+import { shortString } from '@/utils/string'
 import { useQuery } from '@apollo/client'
 import { sendGAEvent } from '@next/third-parties/google'
+import type { SortingState } from '@tanstack/react-table'
+import { CopyButton } from 'components/common/CopyButton'
 import { useEvmExplorerBanner } from 'components/common/EvmExplorerBanner'
-import { ExportButton } from 'components/common/ExportButton'
-import { Pagination } from 'components/common/Pagination'
 import { SearchBar } from 'components/common/SearchBar'
+import { SortedTable } from 'components/common/SortedTable'
 import { Spinner } from 'components/common/Spinner'
+import { StatusIcon } from 'components/common/StatusIcon'
+import { INTERNAL_ROUTES } from 'constants/routes'
+import dayjs from 'dayjs'
+import relativeTime from 'dayjs/plugin/relativeTime'
 import { Extrinsic, ExtrinsicWhereInput, ExtrinsicsConnectionQuery } from 'gql/graphql'
-import { FC, useCallback, useEffect, useMemo, useState } from 'react'
+import useDomains from 'hooks/useDomains'
+import Link from 'next/link'
+import { FC, useEffect, useMemo, useState } from 'react'
 import { useErrorHandler } from 'react-error-boundary'
+import type { Cell } from 'types/table'
 import { NotFound } from '../layout/NotFound'
 import { ExtrinsicListFilter } from './ExtrinsicListFilter'
-import { ExtrinsicTable } from './ExtrinsicTable'
 import { QUERY_EXTRINSIC_LIST_CONNECTION } from './query'
 
+dayjs.extend(relativeTime)
+
 export const ExtrinsicList: FC = () => {
-  const [currentPage, setCurrentPage] = useState(0)
-  const [lastCursor, setLastCursor] = useState<string | undefined>(undefined)
+  const { selectedChain, selectedDomain } = useDomains()
+  const [sorting, setSorting] = useState<SortingState>([{ id: 'id', desc: false }])
+  const [pagination, setPagination] = useState({
+    pageSize: PAGE_SIZE,
+    pageIndex: 0,
+  })
   const [filters, setFilters] = useState<ExtrinsicWhereInput>({})
   const novaExplorerBanner = useEvmExplorerBanner('txs')
+
+  const variables = useMemo(
+    () => ({
+      first: pagination.pageSize,
+      after:
+        pagination.pageIndex > 0
+          ? (pagination.pageIndex * pagination.pageSize).toString()
+          : undefined,
+    }),
+    [pagination.pageSize, pagination.pageIndex],
+  )
 
   const { data, error, loading } = useQuery<ExtrinsicsConnectionQuery>(
     QUERY_EXTRINSIC_LIST_CONNECTION,
     {
-      variables: { first: PAGE_SIZE, after: lastCursor, where: filters },
+      variables,
       pollInterval: 6000,
     },
   )
@@ -44,37 +69,84 @@ export const ExtrinsicList: FC = () => {
     () => extrinsicsConnection && extrinsicsConnection.totalCount,
     [extrinsicsConnection],
   )
-  const totalLabel = useMemo(() => numberWithCommas(Number(totalCount)), [totalCount])
-  const pageInfo = useMemo(
-    () => extrinsicsConnection && extrinsicsConnection.pageInfo,
-    [extrinsicsConnection],
+  const pageCount = useMemo(
+    () => Math.ceil(Number(totalCount) / pagination.pageSize),
+    [totalCount, pagination],
   )
+  const totalLabel = useMemo(() => numberWithCommas(Number(totalCount)), [totalCount])
   const modules = useMemo(
     () => data && data.extrinsicModuleNames.map((module) => module.name.split('.')[0]),
     [data],
   )
 
-  const handleNextPage = useCallback(() => {
-    if (!pageInfo) return
-    setCurrentPage((prev) => prev + 1)
-    setLastCursor(pageInfo.endCursor)
-  }, [pageInfo])
-
-  const handlePreviousPage = useCallback(() => {
-    if (!pageInfo) return
-    setCurrentPage((prev) => prev - 1)
-    setLastCursor(pageInfo.endCursor)
-  }, [pageInfo])
-
-  const onChange = useCallback((page: number) => {
-    setCurrentPage(Number(page))
-
-    const newCount = page > 0 ? PAGE_SIZE * Number(page + 1) : PAGE_SIZE
-    const endCursor = newCount - PAGE_SIZE
-
-    if (endCursor === 0 || endCursor < 0) return setLastCursor(undefined)
-    setLastCursor(endCursor.toString())
-  }, [])
+  const columns = useMemo(
+    () => [
+      {
+        accessorKey: 'block',
+        header: 'Extrinsic Id',
+        enableSorting: false,
+        cell: ({ row }: Cell<Extrinsic>) => (
+          <Link
+            key={`${row.index}-extrinsic-block`}
+            className='hover:text-purpleAccent'
+            href={INTERNAL_ROUTES.extrinsics.id.page(
+              selectedChain.urls.page,
+              selectedDomain,
+              row.original.id,
+            )}
+          >
+            <div>{`${row.original.block.height}-${row.index}`}</div>
+          </Link>
+        ),
+      },
+      {
+        accessorKey: 'timestamp',
+        header: 'Time',
+        enableSorting: false,
+        cell: ({ row }: Cell<Extrinsic>) => (
+          <div key={`${row.index}-extrinsic-time`}>
+            {dayjs(row.original.block.timestamp).fromNow(true)}
+          </div>
+        ),
+      },
+      {
+        accessorKey: 'status',
+        header: 'Status',
+        enableSorting: false,
+        cell: ({ row }: Cell<Extrinsic>) => (
+          <div
+            className='md:flex md:items-center md:justify-start md:pl-5'
+            key={`${row.index}-home-extrinsic-status`}
+          >
+            <StatusIcon status={row.original.success} />
+          </div>
+        ),
+      },
+      {
+        accessorKey: 'action',
+        header: 'Action',
+        enableSorting: false,
+        cell: ({ row }: Cell<Extrinsic>) => (
+          <div key={`${row.index}-extrinsic-action`}>
+            {row.original.name.split('.')[1].toUpperCase()}
+          </div>
+        ),
+      },
+      {
+        accessorKey: 'blockhash',
+        header: 'Block hash',
+        enableSorting: false,
+        cell: ({ row }: Cell<Extrinsic>) => (
+          <div key={`${row.index}-extrinsic-hash}`}>
+            <CopyButton value={row.original.hash} message='Hash copied'>
+              {shortString(row.original.hash)}
+            </CopyButton>
+          </div>
+        ),
+      },
+    ],
+    [selectedChain.urls.page, selectedDomain],
+  )
 
   useEffect(() => {
     try {
@@ -106,21 +178,20 @@ export const ExtrinsicList: FC = () => {
         />
       </div>
       <div className='mt-8 flex w-full flex-col sm:mt-0'>
-        <ExtrinsicTable extrinsics={extrinsics} />
-        <div className='flex w-full justify-between gap-2'>
-          <ExportButton data={extrinsics} filename='extrinsic-list' />
-          {totalCount && pageInfo && (
-            <Pagination
-              nextPage={handleNextPage}
-              previousPage={handlePreviousPage}
-              currentPage={currentPage}
-              pageSize={PAGE_SIZE}
-              totalCount={totalCount}
-              hasNextPage={pageInfo.hasNextPage}
-              hasPreviousPage={pageInfo.hasPreviousPage}
-              onChange={onChange}
+        <div className='w-full'>
+          <div className='my-6 rounded'>
+            <SortedTable
+              data={extrinsics}
+              columns={columns}
+              showNavigation={true}
+              sorting={sorting}
+              onSortingChange={setSorting}
+              pagination={pagination}
+              pageCount={pageCount}
+              onPaginationChange={setPagination}
+              filename='extrinsics-list'
             />
-          )}
+          </div>
         </div>
       </div>
     </div>
