+++ explorer/src/components/Home/HomeExtrinsicList.tsx
@@ -1,10 +1,8 @@
 'use client'
 
-import { PAGE_SIZE } from '@/constants/general'
 import { shortString } from '@/utils/string'
 import { ApolloError } from '@apollo/client'
 import { ArrowLongRightIcon } from '@heroicons/react/24/outline'
-import type { SortingState } from '@tanstack/react-table'
 import { SortedTable } from 'components/common/SortedTable'
 import { StatusIcon } from 'components/common/StatusIcon'
 import { INTERNAL_ROUTES } from 'constants/routes'
@@ -13,7 +11,7 @@ import relativeTime from 'dayjs/plugin/relativeTime'
 import { Extrinsic } from 'gql/graphql'
 import useDomains from 'hooks/useDomains'
 import Link from 'next/link'
-import { FC, useMemo, useState } from 'react'
+import { FC, useMemo } from 'react'
 import type { Cell } from 'types/table'
 import type { HomeQueryDomainQuery, HomeQueryQuery } from '../gql/graphql'
 
@@ -25,28 +23,8 @@ interface HomeExtrinsicListProps {
   data: HomeQueryQuery | HomeQueryDomainQuery
 }
 
-export const HomeExtrinsicListHeader = () => (
-  <div className='mb-6 inline-flex w-full items-center justify-between align-middle'>
-    <div className='text-md uppercase leading-normal text-gray-600 dark:text-white'>
-      Latest Extrinsics
-    </div>
-    <Link
-      data-testid='testLinkExtrinsics'
-      href={INTERNAL_ROUTES.extrinsics.list}
-      className='p-2 transition duration-150 ease-in-out'
-    >
-      <ArrowLongRightIcon stroke='#DE67E4' className='size-6' />
-    </Link>
-  </div>
-)
-
 export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
   const { selectedChain, selectedDomain } = useDomains()
-  const [sorting, setSorting] = useState<SortingState>([{ id: 'id', desc: false }])
-  const [pagination, setPagination] = useState({
-    pageSize: PAGE_SIZE,
-    pageIndex: 0,
-  })
 
   const extrinsics = useMemo(() => data.extrinsics as Extrinsic[], [data.extrinsics])
 
@@ -55,7 +33,6 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
       {
         accessorKey: 'hash',
         header: 'Hash',
-        enableSorting: true,
         cell: ({ row }: Cell<Extrinsic>) => (
           <Link
             className='hover:text-purpleAccent'
@@ -73,7 +50,6 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
       {
         accessorKey: 'block',
         header: 'Block',
-        enableSorting: true,
         cell: ({ row }: Cell<Extrinsic>) => (
           <div key={`${row.index}-home-extrinsic-block`}>{row.original.block.height}</div>
         ),
@@ -81,7 +57,6 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
       {
         accessorKey: 'name',
         header: 'Call',
-        enableSorting: true,
         cell: ({ row }: Cell<Extrinsic>) => (
           <div key={`${row.index}-home-extrinsic-action`}>
             {row.original.name.split('.')[1].toUpperCase()}
@@ -91,7 +66,6 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
       {
         accessorKey: 'timestamp',
         header: 'Time',
-        enableSorting: true,
         cell: ({ row }: Cell<Extrinsic>) => (
           <div key={`${row.index}-home-extrinsic-time`}>
             {dayjs(row.original.timestamp).fromNow(true)} ago
@@ -101,7 +75,6 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
       {
         accessorKey: 'name',
         header: 'Status',
-        enableSorting: true,
         cell: ({ row }: Cell<Extrinsic>) => (
           <div
             className='flex items-center justify-center'
@@ -115,24 +88,32 @@ export const HomeExtrinsicList: FC<HomeExtrinsicListProps> = ({ data }) => {
     [selectedChain.urls.page, selectedDomain],
   )
 
-  const totalCount = useMemo(() => (extrinsics ? extrinsics.length : 0), [extrinsics])
-  const pageCount = useMemo(
-    () => Math.floor(totalCount / pagination.pageSize),
-    [totalCount, pagination],
-  )
-
   return (
     <div className='w-full flex-col rounded-[20px] border border-gray-200 bg-white p-4 dark:border-none dark:bg-gradient-to-r dark:from-purpleUndertone dark:to-blueMedium'>
-      <HomeExtrinsicListHeader />
+      <div className='mb-6 inline-flex w-full items-center justify-between align-middle'>
+        <div className='text-md uppercase leading-normal text-gray-600 dark:text-white'>
+          Latest Extrinsics
+        </div>
+        <Link
+          data-testid='testLinkExtrinsics'
+          href={
+            '/' +
+            selectedChain.urls.page +
+            '/' +
+            selectedDomain +
+            '/' +
+            INTERNAL_ROUTES.extrinsics.list
+          }
+          className='p-2 transition duration-150 ease-in-out'
+        >
+          <ArrowLongRightIcon stroke='#DE67E4' className='size-6' />
+        </Link>
+      </div>
       <SortedTable
         data={extrinsics}
         columns={columns}
-        showNavigation={true}
-        sorting={sorting}
-        onSortingChange={setSorting}
-        pagination={pagination}
-        pageCount={pageCount}
-        onPaginationChange={setPagination}
+        showNavigation={false}
+        pageCount={1}
         filename='home-latest-extrinsics'
       />
     </div>
